refactor(box): move dir of util funcs
